tests/requirements.txt
pyyaml>=6.0.1
pytest>=7.4.0
pytest-cov>=4.1.0
mock>=5.1.0
pycups>=2.0.1
pytest==7.4.3
pytest-cov==4.1.0
flask==3.0.0
flask-cors==4.0.0
pyjwt==2.8.0
werkzeug==3.0.1
requests==2.31.0 